🐛 Fix: 관리자 - 페이지당 유저가 아닌 총 유저 수를 반환하도록 수정

+++ src/main/java/com/poppin/poppinserver/dto/user/response/UserListDto.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Builder
 public record UserListDto(
         List<UserAdministrationDto> userList,
-        int userCnt
+        Long userCnt
 ) {
 }
